DataCollector/requirements.txt
flask
requests
mysql-connector-python
apscheduler
grpcio
grpcio-tools
